Fixed bug #40275 - The sql query in acp_users.php lacks a condition Authorised by: acydburn
git-svn-id: file:///svn/phpbb/branches/phpBB-3_0_0@9310 89ea8834-ac86-4346-8a33-228a782c2dd0

+++ phpBB/includes/acp/acp_users.php
@@ -972,6 +972,7 @@ class acp_users
 					{
 						$sql = 'DELETE FROM ' . LOG_TABLE . '
 							WHERE log_type = ' . LOG_USERS . "
+							AND reportee_id = $user_id
 							$where_sql";
 						$db->sql_query($sql);
